+++ R/game_starts_helpers.R
@@ -1,3 +1,17 @@
+# 24 tiles arranged 5x5 with hole in middle, center of bottom-left tile at (x0, y0)
+df_donut_tiles <- function(seed = NULL, tiles = NULL, x0 = 1.5, y0 = 1.5, face = TRUE) {
+    set.seed(seed)
+    df_txy <- tibble(piece_side = ifelse(face, "tile_face", "tile_back"),
+                     x = x0 + c(rep(seq(0, 8, 2), 2), 0, 2, 6, 8, rep(seq(0, 8, 2), 2)),
+                     y = y0 + c(rep(8, 5), rep(6, 5), rep(4, 4), rep(2, 5), rep(0, 5)))
+    if (is.null(tiles)) {
+        df_tsr <- tibble(suit = rep(1:4, each = 6), rank = rep(1:6, 4))[sample.int(24), ]
+    } else {
+        df_tsr <- process_tiles(tiles)
+    }
+    bind_cols(df_txy, df_tsr)
+}
+
 process_angles <- function(angles) {
     angles <- gsub("\\^", "0", angles)
     angles <- gsub("<", "90", angles)